% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addsteps.R
\name{addsteps}
\alias{addsteps}
\title{Add multiple steps in map}
\usage{
addsteps(
  map = map,
  n = 7,
  m = NULL,
  s = NULL,
  cn = NULL,
  cp = NULL,
  im = NULL,
  tE = NULL,
  tI = NULL,
  tA = NULL,
  ir = NULL,
  cfr = NULL,
  verbose = FALSE
)
}
\arguments{
\item{map}{an epidmap object}

\item{n}{number of steps}

\item{m}{moving \%}

\item{s}{stepsize}

\item{cn}{mean of contacts inside a tile}

\item{cp}{mean of number of people for each contact}

\item{im}{mean of number of people infected}

\item{tE}{number of days of E condition}

\item{tI}{number of days of I condition}

\item{tA}{number of days of A condition}

\item{ir}{infected rate I/(A+I)}

\item{cfr}{case fatality rate}

\item{verbose}{verbose}
}
\value{
an epidmap object
}
\description{
Add multiple steps in map
}
