% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_map.R
\name{animate_map}
\alias{animate_map}
\title{Animate map}
\usage{
animate_map(
  map,
  height = 800,
  width = 800,
  res = 150,
  fps = 10,
  spd = 2,
  end_pause = 30,
  renderer = gganimate::gifski_renderer()
)
}
\arguments{
\item{map}{map object}

\item{height}{height of the map}

\item{width}{width of the map}

\item{res}{resolution}

\item{fps}{frame per second}

\item{spd}{spd}

\item{end_pause}{end_pause}

\item{renderer}{gganimate renderer}
}
\value{
animation
}
\description{
Animate map
}
