% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generate map}
\usage{
generate(
  n = 25,
  P = 1000,
  type = "polygon",
  p = 1,
  rho = 0,
  verbose = T,
  save_movements = F
)
}
\arguments{
\item{n}{A positive integer. Grid dimension}

\item{P}{Total number of individuals}

\item{type}{select type}

\item{p}{A positive integer. People density for each tile - min/max - uniform distribution}

\item{rho}{Spatial autocorrelation parameter}

\item{verbose}{A logical}

\item{save_movements}{save movements in memory}
}
\value{
An epidmap object
}
\description{
Generate an epidmap object with population distributed into spatial units laid on a n-by-n regular squared lattice grid.
}
\details{
\subsection{Map generation}{

For the map generation we considered a population distributed into spatial units laid on a n-by-n regular squared lattice grid. Each square of the grid contains a number of individuals randomly drawn from a uniform distribution ranging between the two values contained in the vector p. This geographical representation is very general in that the map thus generated can represent, e. g., a city divided into blocks or a region divided into smaller spatial union or any other meaningful geographical partition.
}
}
\examples{
\dontrun{
set.seed(12345)
map <- genspmap(n=5, p=c(10, 50))
print(map)
}
}
