% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_attraction.R
\name{move_attraction}
\alias{move_attraction}
\title{Move to attraction}
\usage{
move_attraction(map, m, a = NULL)
}
\arguments{
\item{map}{a map object}

\item{m}{\% of moving people}

\item{a}{a matrix}
}
\value{
map object
}
\description{
Move to attraction
}
