% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_back.R
\name{move_back}
\alias{move_back}
\title{Move back}
\usage{
move_back(map)
}
\arguments{
\item{map}{an epidmap object}
}
\value{
an epidmap object
}
\description{
'move_back' generates the movements to restore the original position of the individuals
after some movements
}
