% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_uniform.R
\name{move_uniform}
\alias{move_uniform}
\title{Move uniform}
\usage{
move_uniform(map, m, s)
}
\arguments{
\item{map}{map object}

\item{m}{\% of moving people}

\item{s}{max step}
}
\value{
map
}
\description{
Move uniform
}
