% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.epidmap.R
\name{plot.epidmap}
\alias{plot.epidmap}
\title{Plot epidmap}
\usage{
\method{plot}{epidmap}(x, ...)
}
\arguments{
\item{x}{an epidmap object}

\item{...}{others}
}
\value{
plot
}
\description{
Plot epidmap
}
