% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.epidmap.R
\name{print.epidmap}
\alias{print.epidmap}
\title{Print epidmap}
\usage{
\method{print}{epidmap}(x, ...)
}
\arguments{
\item{x}{An epidmap object}

\item{...}{others}
}
\value{
print
}
\description{
Print epidmap
}
\examples{
\dontrun{
map <- genmap(n=5, p=c(10, 50))
print(map)
}


}
