% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run simulation}
\usage{
run(
  map,
  daily,
  days = 1,
  tE = 5,
  tA = 14,
  tI = 14,
  ir = 1,
  cfr = 0.15,
  verbose = F
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Run simulation
}
