% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addsteps.R
\name{addsteps}
\alias{addsteps}
\title{Add multiple steps}
\usage{
addsteps(
  map = map,
  n = 7,
  m = NULL,
  s = NULL,
  cn = NULL,
  cp = NULL,
  im = NULL,
  tE = NULL,
  tI = NULL,
  tA = NULL,
  ir = NULL,
  cfr = NULL,
  verbose = FALSE
)
}
\arguments{
\item{map}{An epidmap object}

\item{n}{A positive integer. Number of steps}

\item{m}{A positive float. \% of moving people}

\item{s}{A positive integer. Step size}

\item{cn}{A positive integer. Mean of contacts inside a tile}

\item{cp}{A positive integer. Mean of number of people for each contact}

\item{im}{A positive integer. Mean of number of people infected}

\item{tE}{A positive integer. Number of days of E condition}

\item{tI}{A positive integer. Number of days of I condition}

\item{tA}{A positive integer. Number of days of A condition}

\item{ir}{A positive float. Infected rate I/(A+I)}

\item{cfr}{A positive float.case Fatality rate}

\item{verbose}{A logical}
}
\value{
an epidmap object
}
\description{
Simulate multiple step (i.e. days) of mobility, social interaction and the evolution of the epidemics with the infection and evolution of the disease for each individual in the map.
}
\details{
\subsection{Mobility}{

The contagion mechanism is favoured by people mobility. In this simulation, we assumed that in any moment of time a certain percentage \strong{m} of the population can move between the squares. In this way it is possible to distinguish different epidemic phases such as free-to-move period and lockdown. The commuting during the lockdown period is not only limited by the number of people who move, but also by the extent of their movements. This is a further simulation parameter which is generated by a uniform distribution ranging from \strong{-s} to \strong{s}.
}

\subsection{Social interaction}{

Given the mobility pattern described above, contagion is determined by the social interaction and the contact opportunities. The number of contacts in each square of the grid is assumed to be determined by a random number drawn from a Poisson distribution with parameter, say \strong{cn}, while the number of people involved in the movements is also a Poisson number characterised but a different parameter \strong{cp}. Given these assumptions, a contagion occurs in the following way. If in a meeting it is present at least one asymptomatic or an exposed person, \strong{im} susceptibles will be infected moving in the status of the Exposed.
}

\subsection{Evolution of the epidemics}{

First of all, in order to simulate an artificial population describing the time evolution of an epidemics, we considered a popular model constituted by a system of six differential equations which, in each moment of time, describe six categories of individuals, namely: the susceptibles (S), those exposed to the virus (E), the infected with symptoms (I), those without symptoms (A) and those that are removed from population either because healed (R) or dead (D). This modelling framework is due to the seminal contribution of Hamer (1906), Kermack and McKendrick (1927) and Soper (1929) and it is often referred to as the “SIR model” from the initials of the categories considered in the first simplified formulation: Susceptibles, Infected and Removed.
For the data random generation, we assumed that, if infected, a susceptible element of the population (S) will remain in the exposed state (E) for the time \strong{tE}. After that period the subject can become either infected with symptoms (I) with probability \strong{ir} or without (asymptomatic; symbol A) with probability \strong{1-ir}. The asymptomatic will remain infected (and so still able to transmit the virus) for \strong{tA} days. After this period all the asymptomatic will be considered healed and will pass to the category removed (R). In contrast, the infected people showing symptoms will be healed with probability \strong{1-cfr} or die (D) with probability \strong{cfr} (case death rate).
}
}
\examples{
\dontrun{
set.seed(12345)
map <- genmap(n=5, p=c(10, 50))
map <- map \%>\%
 addsteps(n=7*3, m=0.03, s=2,
          cn=4, cp=2,
          im=3, tE=5, tA=14,
          tI=14, ir=1, cfr=0.15) \%>\%
 addsteps(n=19, m=0.01, s=1,
          cn=1, cp=2,
          im=3, tE=5, tA=14,
          tI=14, ir=1, cfr=0.15)
}
}
