% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmap.R
\name{genmap}
\alias{genmap}
\title{Generate map}
\usage{
genmap(n = 5, p = c(10, 50), verbose = T)
}
\arguments{
\item{n}{grid dimension}

\item{p}{uniform distribution of people}

\item{verbose}{verbose?}
}
\value{
An epidmap object
}
\description{
Generate map
}
\examples{
genmap(n=5, p=c(10, 50))

}
