% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meet.R
\name{meet}
\alias{meet}
\title{Meet}
\usage{
meet(map, cn, cp, im, parallel = F)
}
\arguments{
\item{map}{a map object}

\item{cn}{number of contacts in each cell}

\item{cp}{number of people in each contact}

\item{im}{number of infected}

\item{parallel}{parallel computation}
}
\value{
a map object
}
\description{
Meet
}
