% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run simulation}
\usage{
run(
  map,
  daily,
  days = 1,
  tE = 5,
  tA = 14,
  tI = 14,
  ir = 1,
  cfr = 0.15,
  verbose = F
)
}
\arguments{
\item{map}{a map object}

\item{daily}{steps of functions}

\item{days}{number of days}

\item{tE}{A positive integer. Number of days of E condition}

\item{tA}{A positive integer. Number of days of A condition}

\item{tI}{A positive integer. Number of days of I condition}

\item{ir}{A positive float. Infected rate I/(A+I)}

\item{cfr}{A positive float.case Fatality rate}

\item{verbose}{A logical}
}
\value{
an epidmap object
}
\description{
Run simulation
}
